change recipe id column to string
