--- alembic/versions/643184784d8c_add_new_model_definitions.py
@@ -1,152 +1,0 @@
-"""add new model definitions
-
-Revision ID: 643184784d8c
-Revises: 969a5b9b30fb
-Create Date: 2023-05-10 00:57:08.115625
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '643184784d8c'
-down_revision = '969a5b9b30fb'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('ingredients',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('image_url', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_ingredients_name'), 'ingredients', ['name'], unique=False)
-    op.create_table('tags',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_tags_name'), 'tags', ['name'], unique=False)
-    op.create_table('user',
-    sa.Column('first_name', sa.String(), nullable=False),
-    sa.Column('last_name', sa.String(), nullable=False),
-    sa.Column('username', sa.String(), nullable=False),
-    sa.Column('email', sa.String(), nullable=True),
-    sa.Column('password_hash', sa.String(), nullable=False),
-    sa.Column('role', sa.String(), server_default='user', nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('username')
-    )
-    op.create_index(op.f('ix_user_email'), 'user', ['email'], unique=True)
-    op.create_table('comments',
-    sa.Column('comment', sa.String(), nullable=True),
-    sa.Column('recipe_id', sa.String(), nullable=False),
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['recipe_id'], ['recipes.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('nutritional_info',
-    sa.Column('calories', sa.Float(), nullable=False),
-    sa.Column('carbohydrates', sa.Float(), nullable=False),
-    sa.Column('fat', sa.Float(), nullable=False),
-    sa.Column('protein', sa.Float(), nullable=False),
-    sa.Column('fiber', sa.Float(), nullable=False),
-    sa.Column('sugar', sa.Float(), nullable=False),
-    sa.Column('sodium', sa.Float(), nullable=False),
-    sa.Column('recipe_id', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['recipe_id'], ['recipes.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('ratings',
-    sa.Column('recipe_id', sa.String(), nullable=False),
-    sa.Column('user_id', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['recipe_id'], ['recipes.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('recipe_ingredients',
-    sa.Column('recipe_id', sa.String(), nullable=False),
-    sa.Column('ingredient_id', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['ingredient_id'], ['ingredients.id'], ),
-    sa.ForeignKeyConstraint(['recipe_id'], ['recipes.id'], ),
-    sa.PrimaryKeyConstraint('recipe_id', 'ingredient_id', 'id')
-    )
-    op.create_table('recipe_tags',
-    sa.Column('recipe_id', sa.String(), nullable=False),
-    sa.Column('tag_id', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['recipe_id'], ['recipes.id'], ),
-    sa.ForeignKeyConstraint(['tag_id'], ['tags.id'], ),
-    sa.PrimaryKeyConstraint('recipe_id', 'tag_id', 'id')
-    )
-    op.add_column('recipes', sa.Column('description', sa.Text(), nullable=False))
-    op.add_column('recipes', sa.Column('prep_time', sa.Integer(), nullable=False))
-    op.add_column('recipes', sa.Column('cooking_time', sa.Integer(), nullable=False))
-    op.add_column('recipes', sa.Column('servings', sa.Integer(), nullable=False))
-    op.add_column('recipes', sa.Column('image_url', sa.String(), nullable=False))
-    op.add_column('recipes', sa.Column('user_id', sa.String(), nullable=False))
-    op.add_column('recipes', sa.Column('created_at', sa.DateTime(), nullable=False))
-    op.add_column('recipes', sa.Column('updated_at', sa.DateTime(), nullable=True))
-    op.drop_index('ix_recipes_id', table_name='recipes')
-    op.create_foreign_key(None, 'recipes', 'user', ['user_id'], ['id'])
-    op.drop_column('recipes', 'serving')
-    op.drop_column('recipes', 'image_file_path')
-    op.drop_column('recipes', 'calories')
-    op.drop_column('recipes', 'ingredients')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('recipes', sa.Column('ingredients', postgresql.ARRAY(sa.VARCHAR()), autoincrement=False, nullable=False))
-    op.add_column('recipes', sa.Column('calories', sa.DOUBLE_PRECISION(precision=53), autoincrement=False, nullable=True))
-    op.add_column('recipes', sa.Column('image_file_path', sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.add_column('recipes', sa.Column('serving', sa.INTEGER(), autoincrement=False, nullable=True))
-    op.drop_constraint(None, 'recipes', type_='foreignkey')
-    op.create_index('ix_recipes_id', 'recipes', ['id'], unique=False)
-    op.drop_column('recipes', 'updated_at')
-    op.drop_column('recipes', 'created_at')
-    op.drop_column('recipes', 'user_id')
-    op.drop_column('recipes', 'image_url')
-    op.drop_column('recipes', 'servings')
-    op.drop_column('recipes', 'cooking_time')
-    op.drop_column('recipes', 'prep_time')
-    op.drop_column('recipes', 'description')
-    op.drop_table('recipe_tags')
-    op.drop_table('recipe_ingredients')
-    op.drop_table('ratings')
-    op.drop_table('nutritional_info')
-    op.drop_table('comments')
-    op.drop_index(op.f('ix_user_email'), table_name='user')
-    op.drop_table('user')
-    op.drop_index(op.f('ix_tags_name'), table_name='tags')
-    op.drop_table('tags')
-    op.drop_index(op.f('ix_ingredients_name'), table_name='ingredients')
-    op.drop_table('ingredients')
-    # ### end Alembic commands ###
